pipeline/requirements.txt
ufcscraper>=1.1.0
supabase>=2.0.0
python-dotenv>=1.0.0
